feat: breaking queries and reusing them
