+++ src/app/file-row.tsx
@@ -1,8 +1,8 @@
 import { FileIcon, Folder as FolderIcon } from "lucide-react"
 import Link from "next/link"
-import type { files, folders } from "~/server/db/schema"
+import type { fileSchema, folderSchema } from "~/server/db/schema"
 
-export function FileRow(props: { file: typeof files.$inferSelect }) {
+export function FileRow(props: { file: typeof fileSchema.$inferSelect }) {
   const { file } = props
   return (
     <li key={file.id} className="px-6 py-4 border-b border-gray-700 hover:bg-gray-750">
@@ -24,7 +24,7 @@ export function FileRow(props: { file: typeof files.$inferSelect }) {
   )
 }
 
-export function FolderRow(props: { folder: typeof folders.$inferSelect }) {
+export function FolderRow(props: { folder: typeof folderSchema.$inferSelect }) {
   const { folder } = props
   return (
     <li key={folder.id} className="px-6 py-4 border-b border-gray-700 hover:bg-gray-750">
