+++ src/server/db/schema.ts
@@ -4,7 +4,7 @@ export const createTable = singlestoreTableCreator(
   (name) => `drive_tutorial_${name}`,
 );
 
-export const files = createTable('files', {
+export const fileSchema = createTable('files', {
   id: bigint('id', { mode: "number", unsigned: true }).primaryKey().autoincrement(),
   name: text('name').notNull(),
   size: int('size').notNull(),
@@ -16,7 +16,7 @@ export const files = createTable('files', {
   ]
 });
 
-export const folders = createTable('folders', {
+export const folderSchema = createTable('folders', {
   id: bigint('id', { mode: "number", unsigned: true }).primaryKey().autoincrement(),
   name: text('name').notNull(),
   parent: bigint('parent', { mode: "number", unsigned: true }),
